web: behaviour: routes are based off pipeline ID
concourse/concourse#5921

Following the previous commit that changed the way we identify pipelines
to use the database ID, the URLs should also be based off the database
IDs. Any route that was `/teams/:team/pipelines/:pipeline` now becomes
`/pipelines/:id`.

This commit also includes all of the test changes that were required to
make this and the previous commit compile.

p.s. a lot of the test changes aren't directly related to the refactor.
Most tests used pipeline id 0 as the first pipeline, which isn't
actually a valid pipeline id in Concourse (they start from id 1)

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Clara Fu <cfu@pivotal.io>