Fly execute: pass worker tags to input.
Fix #5672.

When issuing `fly execute` and selecting a platform
where there are *only* tagged worker the command fails
because DetermineInputs does not pass along the tags
specified in the command and thus, at the end of the chain,
SelectWorkers will return that no worker satisfy the
request (because the are no untagged workers).

This commit passes the worker tags along the chain so that
the input can be set up.

Signed-off-by: Alessandro Degano <alessandro.degano@pix4d.com>