fly: make `--client-cert` flag name consistent
This is a follow-up to #6017, which added a new `--client-cert` flag to
`fly login` and `fly sync` - only, the latter had the flag as
`--client-certificate`, not `--client-cert`.

This was due to a review comment I left that wasn't totally clear, and I
didn't feel it was worth another review cycle to get this changed, so
here we are!

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>