atc: structure: construct step delegate in step
Previously, we were constructing step delegates and passing them into
the step directly. This commit instead passes in a step delegate
factory, and the steps construct the step delegate at runtime.

p.s. there was a delegate factory that existed before, but it was kinda
useless - #6022 got rid of it entirely.

This may seem like a useless change, and for the time being it is. The
motivation for this is to allow the step delegate to be constructed
using information that can only exist at runtime.

In particular, the step delegates currently require vars.BuildVariables
in order to:
1. Expose creds and local variables to the steps
2. Perform secret redaction

In coming commits, I plan to extend the `RunState` that's passed into
every `Step` to also deal with variables, rather than relying on a step
delegate to expose this. This are a few motivations for this:
1. Allow things to be interpolated more easily - a big driving force
   behind this change is to make it possible to interpolate most
   everything in a build plan. This could've been done by injecting a
   step delegate into every step type, but that felt awkward
2. Make the across step construction less awkward - since step delegates
   had to be constructed statically with their local vars at build time,
   we needed some tricky logic in the builder to figure out what to run
   in parallel. This proved to be difficult to test, leading to a lack
   of test coverage.
3. Opens the door for the across step to run with a dynamic list of
   values. This is definitely down the line, and requires other changes
   to make it work, but requiring the variables to be set at build time
   makes this impossible. It also would mean you can't interpolate
   individual values (i.e. not the full list), which *is* something that
   should be possible with the static across step

p.s. this commit removes some test coverage for the across step around
properly setting the variable scopes. I'll add this back in a later
commit

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>