atc: structure: move vars.BuildVariables to exec pkg
Something specific to Builds didn't belong outside of the atc package.
Given that it's only consumed directly by the exec.RunState, it made
sense to go there.

I considered just inlining all of the logic in RunState, but decided
it's better to keep RunState lean, since it already has a few
responsibilities.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>