Fix: ignore unknown keys in garden ini
We started parsing the garden ini as part of #6087. After that PR the
worker would fail to come up if the garden.ini contained any unknown
keys (unknown to Concourse).

This commit will unblock the pipeline. We should spend some time
thinking about this area of the codebase and figure out what the best
strategy moving forward should be.

Signed-off-by: Sameer Vohra <svohra@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>