worker: use default uid if /etc/passwd not found
concourse/concourse#6054

If the username is "root", but we can't find a /etc/passwd to lookup
uid:gid, then assume 0:0. This seems to be the convention among most
linux distributions, and is the assumption that Guardian makes:

It looks like both golang and runc assumes that the passwd file is
located at `/etc/passwd`. If runc fails to read the file, it'll use
whatever defaults are passed in. It happens that Guardian uses 0:0.

Signed-off-by: Rui Yang <ruiya@vmware.com>
Co-authored-by: Bohan Chen <bochen@pivotal.io>