vars: structure: Variable impls traverse fields
Previously, vars.Variable implementations only considered the `Path`
(and occasionally, `Source`) fields, and completely ignored the `Fields`
of the passed in `vars.Reference` - it was up to `vars.Template` to
actually traverse the fields.

This is certainly odd - I would expect as a caller of a `vars.Variables`
that the Fields would be traversed given that I'm passing in a full
Reference.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>