Check that a pipeline contains at least one job
From concourse/concourse#6111 we found that it was possible for any yaml
to pass validation and that ended up having some side-affects that
brought down the ATC. This commit will now ensure that any yaml someone
tries to pass off as a pipeline needs to contain at least one job. A
pipeline without a job is not very useful.

More context in this comment:
https://github.com/concourse/concourse/pull/6116#issuecomment-708449186

Signed-off-by: Taylor Silva <tsilva@pivotal.io>