atc: behavior: consistently handle ImageArtifact
this was special-cased for RunTaskStep - but now we'll need it for all
three methods.

to be honest, this feels pretty wack - I lost a lot of time on this
because the worker image fetching path doesn't actually respect
ImageArtifact - it's all handled in the worker client, because it needs
to look up the artifact across all workers, while the image fetching
path only has access to the chosen worker.

it's all kind of roundabout - we go from a worker.Volume to a
runtime.Artifact and back to a worker.Volume, by treating
(runtime.Artifact).ID as the handle, even though it's not always a
handle.

seems like this should be cleaned up, but that seems like an exercise
for another PR.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>