atc: behavior: enforce image policy in FetchImage
this involves sending 'privileged' in the policy check request, which
meant I had to change FetchImage, which meant I had to change all the
delegates and all the steps, so there's quite a bit of churn.

the policy checking code is at least a little cleaner than the worker
image fetching counterpart - it doesn't have to sneak the team/pipeline
in through context.Context, because it's right there on the build that
the delegate already has!

now that policy enforcement is implemented in the new code path, we can
start to rip out all the worker image fetching code - i'll go do that
now!

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>