atc: behavior: fix resource container platform
previously, image fetching would always happen on the same worker that
we're creating the container on. this resulted in only placing resource
containers on workers which support either their base resource type or
the bottom-most resource type used for fetching the resource type image
- typically `registry-image`.

this naturally led to resource containers only being placed on linux
workers, which are the only workers which support the `registry-image`
resource type (or any resource type at all).

now that image fetching is run as a separate step and has been decoupled
from the container creation logic, it may happen on any worker, and when
it comes time to create the resource container there's technically
nothing specifying what platform to create it on. tasks don't have this
problem because everyone has been dutifully specifying 'platform:' in
the task config - but that's not a thing for resources.

so this commit brings it closer to the old behavior: we'll only run
resource containers on workers which support the bottom-most resource
type used for fetching its image. we'll have to see if this still makes
sense if/when people start to implement resource types on other
platforms.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>