web: behaviour: switch back to old routes
concourse/concourse#5921

At the start of this PR, I had some large commits that switched the
routes from /teams/:team_name/pipelines/:pipeline_name to
/pipelines/:pipeline_id (sim. for the API endpoints), as it's more
convenient to work with just the ID. We ended up not wanting to do that,
because we don't want to get rid of the "friendly" routes.

So, here's a massive commit that switches everything over to the old
routes. The raw instance var JSON is specified in the query string of
the routes for pipeline instances.

Apologies for the massive commit - it's not so easy to split it up into
more managable chunks.

This reverts commit d761d67eda3c45002b422f3c2dea377f6d4c66e1.
This reverts commit 6199e9dae45164d678e913eedc1774541904392f.
This reverts commit d702d80f10a487959d7b9c7cda35fd1cffc951dd.
This reverts commit e7d278e9b68dc9f24cfa5fe878a4fdf789b499d8.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>