topgun: fix volume_gc_test
now that 'get' steps for image fetching are associated to builds, this
test started to fail because it'd find 'get' step and expect its volume
to go away, but image resource cache volumes are kept for 24 hours.

this fix isn't pretty, but it at least makes it test only the volumes
for the container that the test cares about.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>