atc: behavior: only tie resource checks to session
for non-resource checks, just associate the container to the build
instead.

this fixes a topgun test which expected to see the image's 'check'
container in 'fly containers'.

without this change, the check containers will not be visible in `fly
containers` because the containers API only returns containers for check
sessions that are associated to pipeline resources visible to the user.
(ain't that confusing?)

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>