web: behaviour: dot notation for instance vars query params
concourse/concourse#5921

Rather than encoding the raw JSON in the query string (which is not very
readable, since many characters get percent encoded), use dot notation
in the query string. For instance, consider the following JSON object:

{
  "foo": "bar",
  "baz": {"qux": 123}
}

This would be encoded in the query string as:

?var.foo="bar"&var.baz.qux=123

Elm's built-in query parsing module doesn't support parsing dynamic
query params - we worked around this by parsing the query params that
start with "var." manually. This meant we had to convert all of the
route parsing to return a deferred Route (that requires the parsed
instance vars to construct) rather than a Route directly, since we only
know the instance vars after parsing everything else.

Rather than duplicating logic between Elm and Javascript (which is used
on the pipeline page), we simply pass along the query params that start
with "var." to the hrefs for the job and resource pages.

The new DotNotation module contains the logic for parsing/serializing
this form (a.b.c="raw json"), as well as "expanding" a "flattened" form
of JSON objects into a full JSON object. This notion of "expanding" and
"flattening" is taken from 3dc8afd2a475905a6b51475440eb4c8c7d239e4b,
which does the same thing in Go.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Vikram Yadav <vyadav@pivotal.io>